Use HDL main branch & update tools versions
Vivado required version is set to 2023.2
Matlab required version is set to R2023b

+++ synth_designs.sh
@@ -2,21 +2,28 @@
 
 BOARD=$1
 MLFLAGS="-nodisplay -nodesktop -nosplash"
-
 if [ -z "$MLRELEASE" ]
 then
-	MLRELEASE=R2022a
+	MLRELEASE=R2023b
 fi
 
-MLPATH=/usr/local/MATLAB
+MLPATH=/emea/mediadata/opt/MATLAB
 
-cd ../.. 
+if [ -z "$VIVADORELEASE" ]
+then
+        VIVADORELEASE=2023.2
+fi
+
+cd ../..
 cp hdl/vendor/AnalogDevices/hdlcoder_board_customization.m test/hdlcoder_board_customization_local.m
 sed -i "s/hdlcoder_board_customization/hdlcoder_board_customization_local/g" test/hdlcoder_board_customization_local.m
-source /opt/Xilinx/Vivado/2022.2/settings64.sh
+source /emea/mediadata/opt/Xilinx/Vivado/$VIVADORELEASE/settings64.sh
 Xvfb :77 &
 export DISPLAY=:77
 export SWT_GTK3=0
-source /opt/Xilinx/Vivado/2022.2/settings64.sh
-$MLPATH/$MLRELEASE/bin/matlab $MLFLAGS -r "cd('test');runSynthTests('$BOARD');"
-kill -9 `pidof Xvfb`
+source /emea/mediadata/opt/Xilinx/Vivado/$VIVADORELEASE/settings64.sh
+$MLPATH/$MLRELEASE/bin/matlab $MLFLAGS -r "addpath(genpath(pwd));cd('test');runSynthTests('$BOARD');"
+pidof Xvfb
+if [ $? -eq 0 ]; then
+    kill -9 `pidof Xvfb`
+fi
